i feel sick I've coded myself into a black hole and I can't get out... I wrote code which may or may not work, and further - I can't remember what the definition of "work" actually is in this case.
I think I want `createRetrierFn` to take any of an iterator, an array,
a number (from which we create an iterator) a generator, or a regular
funciton (which returns a function or an array, or an iterable). These
would all be called / iterated through internally producing a set of
either numbers or promises. Numbers are treated as an amount of time to
delay the next invokation for, while promises are treated as sleep intervals
(and so the next invokation occurs when they resolve).