Allow optional retry behaviour in retry function.
Sometimes you want to throw an error rather than retrying (e.g. if
wrapping an HTTP requesting function you might want to retry HTTP 5XX
but not 4XX).  It's possible to just not throw an error and return some
sort of error response which can be detected in the caller, but this
results in a more complex flow which can be more easily accommodated
here.

Additionally the retrier function is simplified (no setTimeout) and
adjusted to retry sync errors as well as async.

Tests added to confirm behaviour.